refactor: Notification 생성 전략 패키지 수정 (#71)
