+++ mohaeng/src/main/java/com/mohaeng/notification/domain/model/NotificationMakeStrategy.java
@@ -1,7 +1,6 @@
-package com.mohaeng.notification.application.eventhandler;
+package com.mohaeng.notification.domain.model;
 
 import com.mohaeng.common.notification.NotificationEvent;
-import com.mohaeng.notification.domain.model.Notification;
 
 import java.lang.reflect.ParameterizedType;
 import java.util.List;
@@ -10,7 +9,6 @@ public abstract class NotificationMakeStrategy<T extends NotificationEvent> {
 
     /**
      * 처리할 수 있는 이벤트 클래스
-     * - Overriding 하지 말 것
      * <p>
      * [EX]
      * class Sample extends NotificationMakeStrategy<SampleEvent> {} 인 경우
@@ -21,6 +19,7 @@ public abstract class NotificationMakeStrategy<T extends NotificationEvent> {
         return (Class<T>) (((ParameterizedType) getClass().getGenericSuperclass()).getActualTypeArguments()[0]);
     }
 
+    @SuppressWarnings("unchecked")
     public final List<Notification> make(final NotificationEvent notificationEvent) {
         T event = (T) notificationEvent;
         return makeNotifications(event);
