+++ mohaeng/src/main/java/com/mohaeng/notification/domain/model/strategy/ApplicationProcessedNotificationMakeStrategy.java
@@ -1,7 +1,7 @@
-package com.mohaeng.notification.application.eventhandler.strategy;
+package com.mohaeng.notification.domain.model.strategy;
 
 import com.mohaeng.applicationform.domain.event.ApplicationProcessedEvent;
-import com.mohaeng.notification.application.eventhandler.NotificationMakeStrategy;
+import com.mohaeng.notification.domain.model.NotificationMakeStrategy;
 import com.mohaeng.notification.domain.model.Notification;
 import com.mohaeng.notification.domain.model.kind.ApplicationProcessedNotification;
 import com.mohaeng.notification.domain.model.value.Receiver;
