+++ mohaeng/src/test/java/com/mohaeng/notification/domain/model/strategy/ApplicationProcessedNotificationMakeStrategyTest.java
@@ -1,4 +1,4 @@
-package com.mohaeng.notification.application.eventhandler.strategy;
+package com.mohaeng.notification.domain.model.strategy;
 
 import com.mohaeng.applicationform.domain.event.ApplicationProcessedEvent;
 import com.mohaeng.applicationform.domain.event.ClubJoinApplicationCreatedEvent;
@@ -6,6 +6,7 @@ import com.mohaeng.applicationform.domain.event.OfficerApproveClubJoinApplicatio
 import com.mohaeng.applicationform.domain.event.OfficerRejectClubJoinApplicationEvent;
 import com.mohaeng.common.notification.NotificationEvent;
 import com.mohaeng.notification.domain.model.Notification;
+import com.mohaeng.notification.domain.model.strategy.ApplicationProcessedNotificationMakeStrategy;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -14,40 +15,32 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DisplayName("OfficerApproveClubJoinApplicationNotificationMakeStrategy 는 ")
-class OfficerApproveClubJoinApplicationNotificationMakeStrategyTest {
+@DisplayName("ApplicationProcessedNotificationMakeStrategy 는 ")
+class ApplicationProcessedNotificationMakeStrategyTest {
 
-    private OfficerApproveClubJoinApplicationNotificationMakeStrategy strategy = new OfficerApproveClubJoinApplicationNotificationMakeStrategy();
+    private ApplicationProcessedNotificationMakeStrategy strategy = new ApplicationProcessedNotificationMakeStrategy();
 
     @Test
-    @DisplayName("OfficerApproveClubJoinApplicationEvent 이벤트만을 처리할 수 있다.")
+    @DisplayName("ApplicationProcessedEvent 이벤트만을 처리할 수 있다.")
     void success_test_1() {
         // then
         Assertions.assertAll(
-                () -> assertThat(strategy.supportEvent()).isEqualTo(OfficerApproveClubJoinApplicationEvent.class),
-                () -> assertThat(strategy.supportEvent()).isNotEqualTo(ApplicationProcessedEvent.class),
+                () -> assertThat(strategy.supportEvent()).isEqualTo(ApplicationProcessedEvent.class),
                 () -> assertThat(strategy.supportEvent()).isNotEqualTo(ClubJoinApplicationCreatedEvent.class),
                 () -> assertThat(strategy.supportEvent()).isNotEqualTo(OfficerRejectClubJoinApplicationEvent.class),
+                () -> assertThat(strategy.supportEvent()).isNotEqualTo(OfficerApproveClubJoinApplicationEvent.class),
                 () -> assertThat(strategy.supportEvent()).isNotEqualTo(NotificationEvent.class)
         );
     }
 
     @Test
-    @DisplayName("OfficerApproveClubJoinApplicationEvent 이벤트를 받아 알림을 생성한다.")
+    @DisplayName("ApplicationProcessedEvent 이벤트를 받아 알림을 생성한다.")
     void success_test_2() {
         // given
         Long receiverId = 1L;
 
         // when
-        List<Notification> make = strategy.make(
-                new OfficerApproveClubJoinApplicationEvent(
-                        this,
-                        receiverId,
-                        10L,
-                        11L,
-                        12L,
-                        13L,
-                        14L));
+        List<Notification> make = strategy.make(ApplicationProcessedEvent.reject(this, receiverId, 10L, 11L));
 
         // then
         assertThat(make.size()).isEqualTo(1);
