+++ mohaeng/src/test/java/com/mohaeng/notification/domain/model/strategy/ClubJoinApplicationCreatedNotificationMakeStrategyTest.java
@@ -1,4 +1,4 @@
-package com.mohaeng.notification.application.eventhandler.strategy;
+package com.mohaeng.notification.domain.model.strategy;
 
 import com.mohaeng.applicationform.domain.event.ApplicationProcessedEvent;
 import com.mohaeng.applicationform.domain.event.ClubJoinApplicationCreatedEvent;
@@ -6,6 +6,7 @@ import com.mohaeng.applicationform.domain.event.OfficerApproveClubJoinApplicatio
 import com.mohaeng.applicationform.domain.event.OfficerRejectClubJoinApplicationEvent;
 import com.mohaeng.common.notification.NotificationEvent;
 import com.mohaeng.notification.domain.model.Notification;
+import com.mohaeng.notification.domain.model.strategy.ClubJoinApplicationCreatedNotificationMakeStrategy;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -14,41 +15,34 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DisplayName("OfficerRejectClubJoinApplicationNotificationMakeStrategy 는 ")
-class OfficerRejectClubJoinApplicationNotificationMakeStrategyTest {
+@DisplayName("ClubJoinApplicationCreatedNotificationMakeStrategy 는 ")
+class ClubJoinApplicationCreatedNotificationMakeStrategyTest {
 
-    private OfficerRejectClubJoinApplicationNotificationMakeStrategy strategy = new OfficerRejectClubJoinApplicationNotificationMakeStrategy();
+    private ClubJoinApplicationCreatedNotificationMakeStrategy strategy = new ClubJoinApplicationCreatedNotificationMakeStrategy();
 
     @Test
-    @DisplayName("OfficerRejectClubJoinApplicationEvent 이벤트만을 처리할 수 있다.")
+    @DisplayName("ClubJoinApplicationCreatedEvent 이벤트만을 처리할 수 있다.")
     void success_test_1() {
         // then
         Assertions.assertAll(
-                () -> assertThat(strategy.supportEvent()).isEqualTo(OfficerRejectClubJoinApplicationEvent.class),
-                () -> assertThat(strategy.supportEvent()).isNotEqualTo(ClubJoinApplicationCreatedEvent.class),
+                () -> assertThat(strategy.supportEvent()).isEqualTo(ClubJoinApplicationCreatedEvent.class),
                 () -> assertThat(strategy.supportEvent()).isNotEqualTo(ApplicationProcessedEvent.class),
                 () -> assertThat(strategy.supportEvent()).isNotEqualTo(OfficerApproveClubJoinApplicationEvent.class),
+                () -> assertThat(strategy.supportEvent()).isNotEqualTo(OfficerRejectClubJoinApplicationEvent.class),
                 () -> assertThat(strategy.supportEvent()).isNotEqualTo(NotificationEvent.class)
         );
     }
 
     @Test
-    @DisplayName("OfficerRejectClubJoinApplicationEvent 이벤트를 받아 알림을 생성한다.")
+    @DisplayName("ClubJoinApplicationCreatedEvent 이벤트를 받아 알림을 생성한다.")
     void success_test_2() {
         // given
-        Long receiverId = 1L;
+        List<Long> memberIds = List.of(1L, 2L, 3L, 4L);
 
         // when
-        List<Notification> make = strategy.make(
-                new OfficerRejectClubJoinApplicationEvent(
-                        this,
-                        receiverId,
-                        10L,
-                        11L,
-                        12L,
-                        13L));
+        List<Notification> make = strategy.make(new ClubJoinApplicationCreatedEvent(this, memberIds, 10L, 11L, 12L));
 
         // then
-        assertThat(make.size()).isEqualTo(1);
+        assertThat(make.size()).isEqualTo(memberIds.size());
     }
 }
